seperated video and form
